requirements.txt
chromadb
ollama
aiogram
python-dotenv
